Restructure into src/ directory, extract sub-components, cleanup dead code
- Delete LensViewer-v4.jsx (978 lines of dead code, not imported anywhere)
- Delete PLAN-description-panel-and-deploy.md (completed plan)
- Delete .claude/worktrees/silly-cannon/ (empty old worktree)
- Move all source files into src/ with subdirectories:
  components/, optics/, utils/, content/, lens-data/
- Rename LensViewer-v5.jsx → LensViewer.jsx
- Extract DescriptionPanel, SharedSlidersBar, useMediaQuery, preferences
  from LensViewer.jsx (804→636 lines)
- Update all import paths across source and test files
- Rewrite CLAUDE.md and update README.md to reflect new structure

All 156 tests pass, production build succeeds.
