+++ __tests__/buildLens.test.js
@@ -1,13 +1,13 @@
 import { describe, it, expect } from 'vitest';
-import buildLens, { paraxialTrace } from '../buildLens.js';
-import LENS_DEFAULTS from '../lens-data/defaults.js';
+import buildLens, { paraxialTrace } from '../src/optics/buildLens.js';
+import LENS_DEFAULTS from '../src/lens-data/defaults.js';
 
 /* ── Load all production lens data files ── */
-import ApoLantharRaw  from '../lens-data/ApoLanthar50f2.data.js';
-import NoktonRaw      from '../lens-data/Nokton50f1.data.js';
-import NikkorRaw      from '../lens-data/NikkorZ50mmf18S.data.js';
-import Nikkor105Raw   from '../lens-data/Nikkor105f14E.data.js';
-import Sonnar50f15Raw from '../lens-data/Sonnar50f15.data.js';
+import ApoLantharRaw  from '../src/lens-data/ApoLanthar50f2.data.js';
+import NoktonRaw      from '../src/lens-data/Nokton50f1.data.js';
+import NikkorRaw      from '../src/lens-data/NikkorZ50mmf18S.data.js';
+import Nikkor105Raw   from '../src/lens-data/Nikkor105f14E.data.js';
+import Sonnar50f15Raw from '../src/lens-data/Sonnar50f15.data.js';
 
 const ApoLanthar  = { ...LENS_DEFAULTS, ...ApoLantharRaw };
 const Nokton      = { ...LENS_DEFAULTS, ...NoktonRaw };
