+++ src/components/LensViewer.jsx
@@ -1,13 +1,13 @@
 /**
  * ╔══════════════════════════════════════════════════════════════════════╗
  * ║           INTERACTIVE LENS CROSS-SECTION VISUALIZATION             ║
- * ║                     — Generic Viewer — v5.0                        ║
+ * ║                      — Orchestration Layer —                       ║
  * ╠══════════════════════════════════════════════════════════════════════╣
- * ║  v5 architecture: Diagram rendering extracted to LensDiagramPanel. ║
- * ║  This file is the orchestration layer: top bar, comparison mode,   ║
- * ║  shared controls, view toggles, overlays, and localStorage.        ║
+ * ║  Diagram rendering lives in LensDiagramPanel.  This file handles   ║
+ * ║  top bar, comparison mode, shared controls, view toggles, overlays,║
+ * ║  and localStorage persistence.                                     ║
  * ║                                                                    ║
- * ║  New in v5:                                                        ║
+ * ║  Features:                                                         ║
  * ║    • Side-by-side lens comparison (desktop) / stacked (mobile)     ║
  * ║    • Independent + normalized scale modes                          ║
  * ║    • URL deep links (?a=&b= for comparison, ?lens= for single)    ║
@@ -16,208 +16,34 @@
  */
 
 import { useState, useMemo, useCallback, useEffect, useRef } from "react";
-import ReactMarkdown from "react-markdown";
-import remarkGfm from "remark-gfm";
 
-import T               from './themes.js';
-import buildLens       from './buildLens.js';
-import { LENS_CATALOG, CATALOG_KEYS, mdForKey } from './lensCatalog.js';
+import T               from '../utils/themes.js';
+import buildLens       from '../optics/buildLens.js';
+import { LENS_CATALOG, CATALOG_KEYS, mdForKey } from '../utils/lensCatalog.js';
 import LensDiagramPanel from './LensDiagramPanel.jsx';
-import { parseComparisonParams, buildComparisonURL } from './parseComparisonParams.js';
+import DescriptionPanel from './DescriptionPanel.jsx';
+import SharedSlidersBar from './SharedSlidersBar.jsx';
+import useMediaQuery   from '../utils/useMediaQuery.js';
+import { PREFS_KEY, loadPrefs } from '../utils/preferences.js';
+import { parseComparisonParams, buildComparisonURL } from '../utils/parseComparisonParams.js';
 import { ENABLE_COLOR_TRACING, DEFAULT_COLOR_TRACING,
          ENABLE_DESKTOP_VIEW_TOGGLE, ENABLE_DIAGRAM_ONLY, ENABLE_ANALYSIS_ONLY,
-         ENABLE_COMPARISON, ENABLE_COMPARISON_MOBILE } from './featureFlags.js';
-import { computeFocusPair, computeAperturePair, formatSharedFocusDist, sharedFNumber, snapToCommon } from './comparisonSliders.js';
-import { formatDist } from './optics.js';
+         ENABLE_COMPARISON, ENABLE_COMPARISON_MOBILE } from '../utils/featureFlags.js';
+import { computeFocusPair, computeAperturePair, snapToCommon } from '../utils/comparisonSliders.js';
 import { ErrorDisplay } from './ErrorBoundary.jsx';
-import ABOUT_ME_MD from './AboutMe.md?raw';
-import ABOUT_SITE_MD from './AboutSite.md?raw';
-
-
-/* ── useMediaQuery hook ── */
-function useMediaQuery(query) {
-  const [matches, setMatches] = useState(() =>
-    typeof window !== 'undefined' ? window.matchMedia(query).matches : true
-  );
-  useEffect(() => {
-    const mql = window.matchMedia(query);
-    const handler = (e) => setMatches(e.matches);
-    mql.addEventListener('change', handler);
-    return () => mql.removeEventListener('change', handler);
-  }, [query]);
-  return matches;
-}
-
-/* ── DescriptionPanel — renders markdown with theme-aware styling ── */
-function DescriptionPanel({ markdown, theme: t }) {
-  if (!markdown) {
-    return (
-      <div style={{ padding: 32, color: t.muted, fontSize: 12, fontStyle: "italic" }}>
-        No description available for this lens.
-      </div>
-    );
-  }
-  const components = {
-    h1: ({ children }) => <h1 style={{ fontSize: 16, fontWeight: 700, color: t.descH1, margin: "14px 0 6px", fontFamily: "'DM Sans','Helvetica Neue',sans-serif", letterSpacing: "0.02em" }}>{children}</h1>,
-    h2: ({ children }) => <h2 style={{ fontSize: 13, fontWeight: 600, color: t.descH2, margin: "14px 0 4px", letterSpacing: "0.06em" }}>{children}</h2>,
-    h3: ({ children }) => <h3 style={{ fontSize: 12, fontWeight: 600, color: t.descH2, margin: "10px 0 3px", letterSpacing: "0.04em" }}>{children}</h3>,
-    p: ({ children }) => <p style={{ margin: "6px 0", fontSize: 12, color: t.descText, lineHeight: 1.7 }}>{children}</p>,
-    strong: ({ children }) => <strong style={{ color: t.descH2, fontWeight: 600 }}>{children}</strong>,
-    a: ({ href, children }) => <a href={href} target="_blank" rel="noopener noreferrer" style={{ color: t.descLinkColor, textDecoration: "none", borderBottom: `1px solid ${t.descLinkColor}40` }}>{children}</a>,
-    code: ({ className, children }) => {
-      const isBlock = className?.startsWith('language-');
-      if (isBlock) return <code style={{ fontSize: 11 }}>{children}</code>;
-      return <code style={{ background: t.descCodeBg, padding: "1px 5px", borderRadius: 3, fontSize: 11, color: t.descLinkColor }}>{children}</code>;
-    },
-    pre: ({ children }) => <pre style={{ background: t.descCodeBg, padding: 12, borderRadius: 6, overflow: "auto", margin: "10px 0", fontSize: 11, lineHeight: 1.5 }}>{children}</pre>,
-    ul: ({ children }) => <ul style={{ paddingLeft: 20, margin: "6px 0", fontSize: 12, color: t.descText, lineHeight: 1.7 }}>{children}</ul>,
-    ol: ({ children }) => <ol style={{ paddingLeft: 20, margin: "6px 0", fontSize: 12, color: t.descText, lineHeight: 1.7 }}>{children}</ol>,
-    li: ({ children }) => <li style={{ marginBottom: 3 }}>{children}</li>,
-    blockquote: ({ children }) => <blockquote style={{ borderLeft: `3px solid ${t.descLinkColor}`, paddingLeft: 14, margin: "10px 0", color: t.muted }}>{children}</blockquote>,
-    table: ({ children }) => <div style={{ overflowX: "auto", margin: "10px 0" }}><table style={{ borderCollapse: "collapse", width: "100%", fontSize: 11 }}>{children}</table></div>,
-    thead: ({ children }) => <thead style={{ background: t.descTableHeaderBg }}>{children}</thead>,
-    th: ({ children }) => <th style={{ border: `1px solid ${t.descTableBorder}`, padding: "6px 10px", textAlign: "left", color: t.descH2, fontWeight: 600, fontSize: 10.5 }}>{children}</th>,
-    td: ({ children }) => <td style={{ border: `1px solid ${t.descTableBorder}`, padding: "5px 10px", color: t.descText, fontSize: 10.5 }}>{children}</td>,
-    hr: () => <hr style={{ border: "none", borderTop: `1px solid ${t.descBorder}`, margin: "16px 0" }} />,
-  };
-  return (
-    <div style={{ padding: "16px 24px 24px", fontSize: 12, lineHeight: 1.7 }}>
-      <ReactMarkdown remarkPlugins={[remarkGfm]} components={components}>
-        {markdown}
-      </ReactMarkdown>
-    </div>
-  );
-}
-
-
-/* ── SharedSlidersBar — unified focus/aperture controls for comparison mode ── */
-function SharedSlidersBar({ LA, LB, sharedFocusT, sharedStopdownT, onSharedFocusChange, onSharedStopdownChange, focusPair, aperturePair, theme: t, isWide }) {
-  const { commonPoint: focusCP, minCloseFocus, maxCloseFocus } = focusPair;
-  const { commonPoint: apertureCP, widerFOPEN, narrowerFOPEN, sharedMaxFstop } = aperturePair;
-  const fNum = sharedFNumber(sharedStopdownT, widerFOPEN, sharedMaxFstop);
-  const focusDistStr = formatSharedFocusDist(sharedFocusT, minCloseFocus);
-  const showFocusCP = focusCP > 0.01 && focusCP < 0.99;
-  const showApertureCP = apertureCP > 0.01 && apertureCP < 0.99;
-
-  const sliderWrap = { position: "relative", flex: 1 };
-  const markerStyle = (pos) => ({
-    position: "absolute", left: `${pos * 100}%`, top: -2, transform: "translateX(-50%)",
-    width: 0, height: 0, borderLeft: "4px solid transparent", borderRight: "4px solid transparent",
-    borderTop: `5px solid ${t.sliderAccent}`, opacity: 0.7, pointerEvents: "none",
-  });
-  const markerLineStyle = (pos) => ({
-    position: "absolute", left: `${pos * 100}%`, top: 0, bottom: 0, transform: "translateX(-50%)",
-    width: 1, background: `${t.sliderAccent}40`, pointerEvents: "none",
-  });
-
-  return (
-    <div style={{ padding: isWide ? "14px 24px" : "12px 14px", borderTop: `1px solid ${t.panelBorder}`, background: t.panelBg, transition: "background 0.3s,border-color 0.3s" }}>
-      <div style={{ display: "flex", flexDirection: isWide ? "row" : "column", gap: isWide ? 32 : 16 }}>
-        {/* Focus slider */}
-        <div style={{ flex: 1 }}>
-          <div style={{ display: "flex", alignItems: "center", gap: 10, marginBottom: 8 }}>
-            <span style={{ fontSize: 9.5, color: t.label, letterSpacing: "0.1em", minWidth: 55 }}>FOCUS</span>
-            <span style={{ fontSize: 14, color: t.focusDist, fontWeight: 700, fontVariantNumeric: "tabular-nums" }}>{focusDistStr}</span>
-          </div>
-          <div style={{ display: "flex", alignItems: "center", gap: 8 }}>
-            <span style={{ fontSize: 9, color: t.focusEndpoint }}>{"\u221e"}</span>
-            <div style={sliderWrap}>
-              {showFocusCP && <div style={markerStyle(focusCP)} />}
-              {showFocusCP && <div style={markerLineStyle(focusCP)} />}
-              <input type="range" min="0" max="1" step="0.004" value={sharedFocusT}
-                onChange={e => onSharedFocusChange(parseFloat(e.target.value))}
-                style={{ width: "100%", height: 4, appearance: "none", background: t.sliderTrack, borderRadius: 2, outline: "none", cursor: "pointer", accentColor: t.sliderAccent }} />
-            </div>
-            <span style={{ fontSize: 9, color: t.focusEndpoint }}>{minCloseFocus} m</span>
-          </div>
-          {/* Per-lens readouts */}
-          <div style={{ marginTop: 6, display: "flex", gap: 16, fontSize: 9, color: t.spacingVal, fontVariantNumeric: "tabular-nums" }}>
-            <span>A: {formatDist(focusPair.focusA, LA)}</span>
-            <span>B: {formatDist(focusPair.focusB, LB)}</span>
-          </div>
-        </div>
-
-        {/* Aperture slider */}
-        <div style={{ flex: 1 }}>
-          <div style={{ display: "flex", alignItems: "center", gap: 10, marginBottom: 8 }}>
-            <span style={{ fontSize: 9.5, color: t.label, letterSpacing: "0.1em", minWidth: 75 }}>APERTURE</span>
-            <span style={{ fontSize: 14, color: t.focusDist, fontWeight: 700, fontVariantNumeric: "tabular-nums" }}>
-              f/{fNum < 10 ? fNum.toFixed(1) : Math.round(fNum)}
-            </span>
-          </div>
-          <div style={{ display: "flex", alignItems: "center", gap: 8 }}>
-            <span style={{ fontSize: 9, color: t.focusEndpoint }}>f/{widerFOPEN.toFixed(1)}</span>
-            <div style={sliderWrap}>
-              {showApertureCP && <div style={markerStyle(apertureCP)} />}
-              {showApertureCP && <div style={markerLineStyle(apertureCP)} />}
-              <input type="range" min="0" max="1" step="0.004" value={sharedStopdownT}
-                onChange={e => onSharedStopdownChange(parseFloat(e.target.value))}
-                style={{ width: "100%", height: 4, appearance: "none", background: t.sliderTrack, borderRadius: 2, outline: "none", cursor: "pointer", accentColor: t.sliderAccent }} />
-            </div>
-            <span style={{ fontSize: 9, color: t.focusEndpoint }}>f/{sharedMaxFstop}</span>
-          </div>
-          {/* f-stop quick-select (union of both series) */}
-          <div style={{ marginTop: 6, display: "flex", gap: 14, flexWrap: "wrap", fontSize: 9, color: t.spacingVal, fontVariantNumeric: "tabular-nums" }}>
-            {[...new Set([...LA.fstopSeries, ...LB.fstopSeries])].sort((a, b) => a - b)
-              .filter(n => n >= widerFOPEN - 0.1 && n <= sharedMaxFstop)
-              .map(n => {
-                const stopT = Math.log(n / widerFOPEN) / Math.log(sharedMaxFstop / widerFOPEN);
-                return (
-                  <span key={n}
-                    onClick={() => onSharedStopdownChange(Math.max(0, stopT))}
-                    style={{ cursor: "pointer", opacity: Math.abs(fNum - n) < 0.15 ? 1 : 0.55, transition: "opacity 0.15s" }}>
-                    f/{n}
-                  </span>
-                );
-              })}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-
-/* ── Persistent preferences (localStorage) ── */
-
-const PREFS_KEY = 'lensvis:prefs';
-
-function loadPrefs() {
-  try {
-    const raw = localStorage.getItem(PREFS_KEY);
-    if (!raw) return {};
-    const p = JSON.parse(raw);
-    if (!p || typeof p !== 'object') return {};
-    const out = {};
-    if (typeof p.dark === 'boolean') out.dark = p.dark;
-    if (typeof p.highContrast === 'boolean') out.highContrast = p.highContrast;
-    if (typeof p.showOnAxis === 'boolean') out.showOnAxis = p.showOnAxis;
-    if (typeof p.showOffAxis === 'boolean') out.showOffAxis = p.showOffAxis;
-    if (typeof p.rayTracksF === 'boolean') out.rayTracksF = p.rayTracksF;
-    if (typeof p.showChromatic === 'boolean') out.showChromatic = p.showChromatic;
-    if (typeof p.chromR === 'boolean') out.chromR = p.chromR;
-    if (typeof p.chromG === 'boolean') out.chromG = p.chromG;
-    if (typeof p.chromB === 'boolean') out.chromB = p.chromB;
-    /* v1 compat: lensKey → lensKeyA */
-    const key = p.lensKeyA || p.lensKey;
-    if (typeof key === 'string' && CATALOG_KEYS.includes(key)) out.lensKeyA = key;
-    if (typeof p.lensKeyB === 'string' && CATALOG_KEYS.includes(p.lensKeyB)) out.lensKeyB = p.lensKeyB;
-    if (typeof p.comparing === 'boolean') out.comparing = p.comparing;
-    if (p.scaleMode === 'independent' || p.scaleMode === 'normalized') out.scaleMode = p.scaleMode;
-    return out;
-  } catch { return {}; }
-}
+import ABOUT_ME_MD from '../content/AboutMe.md?raw';
+import ABOUT_SITE_MD from '../content/AboutSite.md?raw';
 
 
 /* =====================================================================
  * §6  RENDERER — React component
  * =====================================================================
- *  v5: Diagram rendering delegated to LensDiagramPanel.  This component
+ *  Diagram rendering delegated to LensDiagramPanel.  This component
  *  handles orchestration, comparison mode, shared controls, and chrome.
  * ------------------------------------------------------------------- */
 
 export default function LensVisualization() {
-  const [prefs] = useState(loadPrefs);
+  const [prefs] = useState(() => loadPrefs(CATALOG_KEYS));
 
   /* ── URL params take precedence over localStorage ── */
   const urlState = useMemo(() => {
