+++ src/lens-data/Nokton50f1.analysis.md
