+++ src/utils/lensCatalog.js
@@ -1,14 +1,13 @@
 /**
- * Lens catalog — auto-registered from ./lens-data/*.data.js
+ * Lens catalog — auto-registered from lens-data/*.data.js
  *
- * Extracted from LensViewer-v4.jsx so both v4 and v5 (and
- * LensDiagramPanel) can share the same catalog without duplicating
- * the import.meta.glob calls.
+ * Shared by LensViewer and LensDiagramPanel so the import.meta.glob
+ * calls are not duplicated.
  */
 
-import LENS_DEFAULTS from './lens-data/defaults.js';
+import LENS_DEFAULTS from '../lens-data/defaults.js';
 
-const _modules = import.meta.glob('./lens-data/*.data.js', { eager: true });
+const _modules = import.meta.glob('../lens-data/*.data.js', { eager: true });
 const LENS_CATALOG = {};
 for (const [path, mod] of Object.entries(_modules)) {
   const data = mod.default;
@@ -20,17 +19,17 @@ for (const [path, mod] of Object.entries(_modules)) {
 }
 const CATALOG_KEYS = Object.keys(LENS_CATALOG).filter(k => LENS_CATALOG[k].visible !== false);
 
-const _mdModules = import.meta.glob('./lens-data/*.analysis.md', { eager: true, query: '?raw', import: 'default' });
+const _mdModules = import.meta.glob('../lens-data/*.analysis.md', { eager: true, query: '?raw', import: 'default' });
 const MD_BY_STEM = {};
 for (const [path, raw] of Object.entries(_mdModules)) {
-  const stem = path.replace('./lens-data/', '').replace('.analysis.md', '');
+  const stem = path.replace('../lens-data/', '').replace('.analysis.md', '');
   MD_BY_STEM[stem] = raw;
 }
 
 /* Map lens key → data file stem so we can find the matching .analysis.md */
 const KEY_TO_STEM = {};
 for (const [path, mod] of Object.entries(_modules)) {
-  const stem = path.replace('./lens-data/', '').replace('.data.js', '').replace('.js', '');
+  const stem = path.replace('../lens-data/', '').replace('.data.js', '').replace('.js', '');
   if (mod.default?.key) KEY_TO_STEM[mod.default.key] = stem;
 }
