+++ src/utils/preferences.js
@@ -1,0 +1,31 @@
+/**
+ * Persistent user preferences via localStorage.
+ */
+
+export const PREFS_KEY = 'lensvis:prefs';
+
+export function loadPrefs(catalogKeys) {
+  try {
+    const raw = localStorage.getItem(PREFS_KEY);
+    if (!raw) return {};
+    const p = JSON.parse(raw);
+    if (!p || typeof p !== 'object') return {};
+    const out = {};
+    if (typeof p.dark === 'boolean') out.dark = p.dark;
+    if (typeof p.highContrast === 'boolean') out.highContrast = p.highContrast;
+    if (typeof p.showOnAxis === 'boolean') out.showOnAxis = p.showOnAxis;
+    if (typeof p.showOffAxis === 'boolean') out.showOffAxis = p.showOffAxis;
+    if (typeof p.rayTracksF === 'boolean') out.rayTracksF = p.rayTracksF;
+    if (typeof p.showChromatic === 'boolean') out.showChromatic = p.showChromatic;
+    if (typeof p.chromR === 'boolean') out.chromR = p.chromR;
+    if (typeof p.chromG === 'boolean') out.chromG = p.chromG;
+    if (typeof p.chromB === 'boolean') out.chromB = p.chromB;
+    /* v1 compat: lensKey → lensKeyA */
+    const key = p.lensKeyA || p.lensKey;
+    if (typeof key === 'string' && catalogKeys.includes(key)) out.lensKeyA = key;
+    if (typeof p.lensKeyB === 'string' && catalogKeys.includes(p.lensKeyB)) out.lensKeyB = p.lensKeyB;
+    if (typeof p.comparing === 'boolean') out.comparing = p.comparing;
+    if (p.scaleMode === 'independent' || p.scaleMode === 'normalized') out.scaleMode = p.scaleMode;
+    return out;
+  } catch { return {}; }
+}
